kullanıcıyla ilişkili postu silme hatası

+++ src/main/java/com/example/facebook/Service/UserService.java
@@ -4,6 +4,8 @@ import com.example.facebook.Dto.CreateUserDto;
 import com.example.facebook.Dto.SubmitPostDto;
 import com.example.facebook.Entity.Post;
 import com.example.facebook.Entity.User;
+import com.example.facebook.Exception.UserNotFoundException;
+import com.example.facebook.Repository.PostRepository;
 import com.example.facebook.Repository.UserRepository;
 import com.jayway.jsonpath.JsonPath;
 import lombok.NoArgsConstructor;
@@ -20,7 +22,9 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final PostRepository postRepository;
     private final ModelMapper modelMapper;
+
 
 
     public List<User> getUsers(){
@@ -41,6 +45,20 @@ public class UserService {
 
 
     public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+        Optional<User> userOptional = userRepository.findById(userId);
+        if(userOptional.isPresent()){
+            User user = userOptional.get();
+            user.getAssignedPosts().remove(user);
+            userRepository.deleteById(userId);
+        }
+        else{
+            throw new UserNotFoundException("Bu kullanıcı mevcut değil." );
+        }
+
+    }
+
+    public User getUserByName(String userName) {
+        return userRepository.findByUserName(userName).orElseThrow(() -> new UserNotFoundException("user not found with this name: " + userName));
+
     }
 }
